requirements.txt
pymongo==3.11.3
paho_mqtt==1.5.1
